refactor: profile flush functions
