+++ packages/platform-sdk-profiles/src/drivers/memory/repositories/setting-repository.ts
@@ -1,7 +1,7 @@
 import { injectable } from "inversify";
 
 import { ISettingRepository } from "../../../contracts";
-import { DataRepository } from "../../../repositories/data-repository";
+import { DataRepository } from "../../../repositories";
 import { emitProfileChanged } from "../helpers";
 
 @injectable()
